install hasura + fix some installation defects

+++ database/uninstall_db.sh
@@ -1,7 +1,13 @@
 #! /bin/sh
 
-NAMESPACE=database
+if [ "$#" -ne "2" ] ; then
+  echo "Usage: $0 release-name namespace"
+  exit 1
+fi
 
-helm uninstall --namespace ${NAMESPACE} postgresql
+RELEASE_NAME=$1
+NAMESPACE=$2
+
+helm uninstall --namespace ${NAMESPACE} ${RELEASE_NAME}
 
 kubectl delete namespace ${NAMESPACE}
